+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Ring fence SQLAlchemy dependency to a facade and integration tests.
 - Add tests for examples.
 - Add _from_str_ and _to_str_ to complement _from_dict_ and _to_dict_ for de-serializing and serializing from JSON.
+- Ring fence jsonschema dependency into a facade.
 
 ## Version 0.12.1 - 2020-01-12